utilities refactor
