+++ server/api/code/lacity_data_api/routers/index.py
@@ -1,7 +1,5 @@
 from fastapi import APIRouter
 
-# from .utilities import build_cache
-
 router = APIRouter()
