+++ server/api/code/lacity_data_api/routers/legacy.py
@@ -4,8 +4,9 @@ from fastapi import APIRouter
 from .api_models import (
     StatusTypes, Filter, Pins, Comparison, Feedback
 )
-from services import status, map, visualizations, requests, comparison, github, email
-from .utilities import build_cache
+from services import (
+    status, map, visualizations, requests, comparison, github, email
+)
 
 router = APIRouter()
 
@@ -19,7 +20,6 @@ use the legacy services code from src as-is.
             description="Provides the status of backend systems")
 async def status_check(status_type: StatusTypes):
     if status_type == StatusTypes.api:
-        await build_cache()
         result = await status.api()
 
     if status_type == StatusTypes.database:
